Converted tests to es6, moved MockLoader to test
This way, there is no danger of MockLoader being shipped with
the src directory.
