+++ test/MockLoader.js
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-import Loader from './Loader';
+import Loader from '../src/Loader';
 
 /**
  * @class
